- initial way labeling - add no-projection option for mapdata - use depth buffer for line clipping to tile - no more scissor, yay! - extract line and poly render function from maprenderer - use one vbo for both polys and lines - use linear interpolator for fling-scroller - add 'exclusive' negative matcher to rendertheme - add some more options to rendertheme, kind of inheritance at least for   lines, see theme - add caching for node tags -> renderinstructions - ...

+++ src/org/mapsforge/android/mapgenerator/MapTile.java
@@ -43,7 +43,7 @@ public class MapTile extends Tile {
 	/**
 	 * distance from center, used in TileScheduler set by updateVisibleList.
 	 */
-	public long distance;
+	public float distance;
 
 	/**
 	 * @param tileX
